+++ src/space.rs
@@ -496,7 +496,7 @@ where
     P: EventProcessor<R>,
 {
     /// Adds a seed to drive the generation of the new rounds model.
-    pub fn seed(&'a mut self, seed: SpaceSeed<R>) -> &'a mut ResetSpaceTrigger<'a, R, P> {
+    pub fn seed(&'a mut self, seed: SpaceSeed<R>) -> &'a mut Self {
         self.seed = Some(seed);
         self
     }