+++ tests/team_test.rs
@@ -23,6 +23,9 @@ struct CustomTeamRules {
 
 impl<R: BattleRules> TeamRules<R> for CustomTeamRules {
     type Id = u32;
+    type Power = EmptyPower;
+    type PowersSeed = ();
+    type PowersAlteration = ();
     type ObjectivesSeed = ();
     type Objectives = ();
 
@@ -513,6 +516,9 @@ fn reset_objectives() {
 
     impl<R: BattleRules> TeamRules<R> for CustomTeamRules {
         type Id = u32;
+        type Power = EmptyPower;
+        type PowersSeed = ();
+        type PowersAlteration = ();
         type ObjectivesSeed = u32;
         type Objectives = u32;
 
@@ -633,6 +639,9 @@ fn check_objectives() {
 
     impl TeamRules<CustomRules> for CustomTeamRules {
         type Id = u32;
+        type Power = EmptyPower;
+        type PowersSeed = ();
+        type PowersAlteration = ();
         type ObjectivesSeed = ();
         type Objectives = ();